+++ examples/snippets/app/examples/suspense-fallbacks/flags.tsx
@@ -3,7 +3,7 @@ import { flag } from '@vercel/flags/next';
 export const hasAuthCookieFlag = flag<boolean>({
   key: 'has-auth-cookie',
   decide({ cookies }) {
-    return cookies.has('ppr-shells-user-id');
+    return cookies.has('suspense-fallbacks-user-id');
   },
 });
